+++ km_api/know_me/models.py
@@ -98,6 +98,16 @@ class EmergencyItem(mixins.IsAuthenticatedMixin, models.Model):
         """
         return self.name
 
+    def get_absolute_url(self):
+        """
+        Get the URL of the instance's detail view.
+
+        Returns:
+            str:
+                The absolute URL of the instance's detail view.
+        """
+        return reverse('know-me:emergency-item-detail', kwargs={'pk': self.pk})
+
     def has_object_read_permission(self, request):
         """
         Check read permissions on the instance for a request.