+++ km_api/know_me/serializers/emergency_item_serializers.py
@@ -7,12 +7,14 @@ from know_me import models
 from know_me.serializers.fields import MediaResourceField
 
 
-class EmergencyItemSerializer(serializers.ModelSerializer):
+class EmergencyItemSerializer(serializers.HyperlinkedModelSerializer):
     """
     Serializer for emergency items.
     """
     media_resource = MediaResourceField(required=False)
+    url = serializers.HyperlinkedIdentityField(
+        view_name='know-me:emergency-item-detail')
 
     class Meta(object):
-        fields = ('id', 'name', 'description', 'media_resource')
+        fields = ('id', 'url', 'name', 'description', 'media_resource')
         model = models.EmergencyItem