+++ km_api/know_me/urls.py
@@ -7,6 +7,7 @@ from know_me import views
 
 
 urlpatterns = [
+    url(r'^emergency-items/(?P<pk>[0-9]+)/$', views.EmergencyItemDetailView.as_view(), name='emergency-item-detail'),   # noqa
     url(r'^media-resources/(?P<pk>[0-9]+)/$', views.MediaResourceDetailView.as_view(), name='media-resource-detail'),   # noqa
     url(r'^items/(?P<pk>[0-9]+)/$', views.ProfileItemDetailView.as_view(), name='profile-item-detail'),                 # noqa
     url(r'^profiles/(?P<pk>[0-9]+)/$', views.ProfileDetailView.as_view(), name='profile-detail'),                       # noqa